jarvis/maria/requirements.txt
mistralai
SpeechRecognition
pyttsx3
pyaudio
pyautogui
python-dotenv
